moved `WavelengthCalibration.add_linear_wavelength_solution` to core as `add_linear_wavelength_solution`
