+++ goodman_pipeline/core/core.py
@@ -142,6 +142,48 @@ def add_wcs_keys(ccd):
     return ccd
 
 
+def add_linear_wavelength_solution(ccd, x_axis, reference_lamp, crpix=1):
+    """Add wavelength solution to the new FITS header
+
+    Defines FITS header keyword values that will represent the wavelength
+    solution in the header so that the image can be read in any other
+    astronomical tool. (e.g. IRAF)
+
+    Args:
+        ccd (CCDData) Instance of :class:`~astropy.nddata.CCDData`
+        x_axis (ndarray): Linearized x-axis in angstrom
+        reference_lamp (str): Name of lamp used to get wavelength solution.
+        crpix (int): reference pixel for defining wavelength solution.
+        Default 1. For most cases 1 should be fine.
+
+
+    Returns:
+        ccd (CCDData) A :class:`~astropy.nddata.CCDData` instance with
+          linear wavelength solution on it.
+
+    """
+    assert crpix > 0
+    new_crpix = crpix
+    new_crval = x_axis[new_crpix - crpix]
+    new_cdelt = x_axis[new_crpix] - x_axis[new_crpix - crpix]
+
+    ccd.header.set('BANDID1', 'spectrum - background none, weights none, '
+                              'clean no')
+    ccd.header.set('WCSDIM', 1)
+    ccd.header.set('CTYPE1', 'LINEAR  ')
+    ccd.header.set('CRVAL1', new_crval)
+    ccd.header.set('CRPIX1', new_crpix)
+    ccd.header.set('CDELT1', new_cdelt)
+    ccd.header.set('CD1_1', new_cdelt)
+    ccd.header.set('LTM1_1', 1.)
+    ccd.header.set('WAT0_001', 'system=equispec')
+    ccd.header.set('WAT1_001', 'wtype=linear label=Wavelength units=angstroms')
+    ccd.header.set('DC-FLAG', 0)
+    ccd.header.set('DCLOG1', 'REFSPEC1 = {:s}'.format(reference_lamp))
+
+    return ccd
+
+
 def bin_reference_data(wavelength, intensity, serial_binning):
     """Bins a 1D array
